pyproject.toml
[tool.poetry]
name = "hw-django"
version = "0.1.0"
description = ""
authors = ["Anastasiia-Eliz <aselizarova@yandex.ru>"]
readme = "README.md"
packages = [{include = "hw_django"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
psycopg2 = "^2.9.3"
Pillow = "^9.2.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.1"
pytest-factoryboy = "^2.5.0"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
